mv KenLmFst class into kenlm.{h,cc}, cosmetic cleanup
