+++ src/lm/kenlm.h
@@ -4,8 +4,17 @@
 #include "lm/model.hh"
 #include "util/murmur_hash.hh"
 
+#include <base/kaldi-common.h>
+#include <fst/fstlib.h>
+#include <fst/fst-decl.h>
+#include <fstext/deterministic-fst.h>
+
 namespace kaldi {
 
+// This class is a thin wrapper of KenLM model,
+// the model itself can be either "trie" or "probing" mode.
+// Kaldi algorithms shouldn't explicitly interact with KenLm class,
+// instead, use provided methods from the fst wrapper (class KenLmFst)
 class KenLm {
  public:
   typedef lm::WordIndex WordIndex;
@@ -15,7 +24,8 @@ class KenLm {
   KenLm() : model_(NULL), vocab_(NULL) { }
 
   ~KenLm() {
-    DELETE(model_);
+    delete model_;
+    model_ = NULL;
     vocab_ = NULL;
     reindex_.clear();
   }
@@ -41,7 +51,12 @@ class KenLm {
   
  private:
   lm::base::Model *model_;
-  const lm::base::Vocabulary* vocab_; // no ownership, it points to models_'s internal vocabuary
+
+  // vocab_ points to models_'s internal vocabulary
+  // no ownership, just for quick reference.
+  const lm::base::Vocabulary* vocab_;
+
+  // Kaldi's output symbol table
   std::unordered_map<std::string, int32> symbol_to_symbol_id_;
 
   // There are really two indexing systems here:
@@ -49,13 +64,91 @@ class KenLm {
   // and KenLM's word index(obtained by hashing the word string).
   // in order to incorperate KenLM into Kaldi,
   // we need to know the mapping between (Kaldi's symbol id -> KenLM's word id)
-  // special symbols:
-  // Kaldi's output symbol table contains two special symbols(<eps> and #0) that do not exist in KenLM,
-  // In any circumstance, these two symbols shouldn't consume any KenLM word,
+  // notes:
+  // <eps> and #0 in symbol table are special,
+  // they do not correspond to any word in KenLM,
+  // Normally, these two symbols shouldn't consume any KenLM word,
   // so the mapping of these two symbols are logically undefined,
-  // we just map them to KenLM's <unk>(which is always indexed as 0) to avoid random invalid mapping.
+  // we just map them to KenLM's <unk>(which is always indexed as 0),
+  // to avoid random invalid mapping.
   std::vector<WordIndex> reindex_; // reindex_[kaldi_symbol_index] -> kenlm word index
 }; // class KenLm
+
+
+//This class wraps KenLM into Kaldi's DeterministicOnDemandFst class,
+//so that Kaldi's fst framework can utilize KenLM as language model
+template<class Arc>
+class KenLmFst : public fst::DeterministicOnDemandFst<Arc> {
+ public:
+  typedef typename Arc::Weight Weight;
+  typedef typename Arc::StateId StateId;
+  typedef typename Arc::Label Label;
+  typedef typename KenLm::State State;
+  typedef typename KenLm::WordIndex WordIndex;
+ 
+  explicit KenLmFst(KenLm *lm)
+   : lm_(lm), num_states_(0), bos_state_id_(0)
+  {
+    MapElem elem;
+    lm->SetStateToBos(&elem.first);
+    elem.second = num_states_;
+
+    std::pair<IterType, bool> ins_result = state_map_.insert(elem);
+    KALDI_ASSERT(ins_result.second == true);
+    state_vec_.push_back(&ins_result.first->first);
+    num_states_++;
+
+    eos_symbol_id_ = lm_->GetSymbolIndex("</s>");
+  }
+  virtual ~KenLmFst() { }
+
+  virtual StateId Start() { 
+    return bos_state_id_;
+  }
+
+  virtual bool GetArc(StateId s, Label label, Arc *oarc) {
+    KALDI_ASSERT(s < static_cast<StateId>(state_vec_.size()));
+    const State* istate = state_vec_[s];
+    MapElem e;
+    WordIndex word = lm_->GetWordIndex(label);
+    BaseFloat log_10_prob = lm_->Score(istate, word, &e.first);
+    e.second = num_states_;
+    std::pair<IterType, bool> result = state_map_.insert(e);
+    if (result.second == true) {
+      state_vec_.push_back(&(result.first->first));
+      num_states_++;
+    }
+
+    oarc->ilabel = label;
+    oarc->olabel = oarc->ilabel;
+    oarc->nextstate = result.first->second;
+    oarc->weight = Weight(-log_10_prob * M_LN10); // KenLM's log10() -> Kaldi's ln()
+
+    return true;
+  }
+
+  virtual Weight Final(StateId s) {
+    Arc oarc;
+    GetArc(s, eos_symbol_id_, &oarc);
+    return oarc.weight;
+  }
+
+ private:
+  KenLm *lm_; // no ownership
+
+ private:
+  typedef std::pair<State, StateId> MapElem;
+  typedef unordered_map<State, StateId, KenLm::StateHasher> MapType;
+  typedef typename MapType::iterator IterType;
+
+  MapType state_map_;
+  std::vector<const State*> state_vec_;
+  StateId num_states_; // state vector index range, [0, num_states_)
+
+  StateId bos_state_id_;  // fst start state id
+  WordIndex eos_symbol_id_;
+
+}; // class KenLmFst
 
 } // namespace kaldi
